pyproject.toml
# -*- coding: utf-8 -*-
[build-system]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "aether"
version = "0.1.0"
description = "业务适配层服务"
authors = ["yancyyu \u003cyancyyu.ok@gmail.com\u003e"]
readme = "README.md"
packages = [{include = "**/*.py"}]  # 包含所有Python文件

[tool.poetry.dependencies]
python = "^3.12"
asyncpg = "^0.30.0"
gunicorn = "^23.0.0"
orjson = "^3.11.5"
python-dotenv = "^1.1.1"
python-multipart = "^0.0.20"
pydantic = "^2.0.0"
pydantic-settings = "^2.0.0"
pydantic-core = "^2.27.0"
sanic = "^25.3.0"
sanic-cors = "^2.2.0"
sanic-ext = "^24.12.0"
uvicorn = "^0.34.2"
httpx = {extras = ["http2"], version = "^0.27.0"}
loguru = "^0.7.0"
redis = "^7.1.0"
tortoise-orm = {extras = ["asyncpg"], version = "^0.25.1"}
cryptography = "^41.0.0"

[tool.ruff]
target-version = "py311"
line-length = 120
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''